refactor: move store to redux
